# Improve OcNodeResourceToInternal and add benchmarks for mashal/unmarshal