Before:
```
go test -benchmem -run=^$ github.com/open-telemetry/opentelemetry-collector/translator/internaldata -bench=.
goos: darwin
goarch: amd64
pkg: github.com/open-telemetry/opentelemetry-collector/translator/internaldata
BenchmarkMetricIntOCToInternal-16          	  521413	      1947 ns/op	    2232 B/op	      37 allocs/op
BenchmarkMetricDoubleOCToInternal-16       	  475279	      2430 ns/op	    2712 B/op	      49 allocs/op
BenchmarkMetricHistogramOCToInternal-16    	  344828	      3263 ns/op	    3672 B/op	      61 allocs/op
BenchmarkMetricSummaryOCToInternal-16      	  519634	      2270 ns/op	    2592 B/op	      43 allocs/op
BenchmarkOcNodeResourceToInternal-16       	  656428	      1823 ns/op	    2580 B/op	      15 allocs/op
BenchmarkInternalResourceToOC-16           	 1365171	       892 ns/op	     950 B/op	       9 allocs/op
```

After:
```
go test -benchmem -run=^$ github.com/open-telemetry/opentelemetry-collector/translator/internaldata -bench=.
goos: darwin
goarch: amd64
pkg: github.com/open-telemetry/opentelemetry-collector/translator/internaldata
BenchmarkMetricIntOCToInternal-16          	  563001	      1858 ns/op	    2136 B/op	      36 allocs/op
BenchmarkMetricDoubleOCToInternal-16       	  497038	      2407 ns/op	    2616 B/op	      48 allocs/op
BenchmarkMetricHistogramOCToInternal-16    	  346826	      3295 ns/op	    3576 B/op	      60 allocs/op
BenchmarkMetricSummaryOCToInternal-16      	  536635	      2228 ns/op	    2496 B/op	      42 allocs/op
BenchmarkOcNodeResourceToInternal-16       	  920998	      1293 ns/op	    1304 B/op	      17 allocs/op
BenchmarkOcResourceNodeUnmarshal-16        	  603808	      1874 ns/op	    1352 B/op	      26 allocs/op
BenchmarkInternalResourceToOC-16           	 1325178	       897 ns/op	     950 B/op	       9 allocs/op
BenchmarkOcResourceNodeMarshal-16          	  497047	      2308 ns/op	    1104 B/op	      33 allocs/op
PASS
ok  	github.com/open-telemetry/opentelemetry-collector/translator/internaldata	11.740s
```

Improvement from `1823 ns/op | 2580 B/op` to `1293 ns/op | 1304 B/op`